Verify mail upon registration (#194)
Before creating a User entity, an Activation token that holds the
users basic information is created. That token is sent to the
specified e-mail address for confirmation. When the user clicks the
link, it is sent via the HTTP Authorization header, triggering the
creation of a User entity and establishing a session.