CatCoder-Import can handle resource-file
Some catcoder-games have additional resource files attached to a challenge (e.g. simulator)
This is done by providing a link to a zip file at the first level. In order to be able to import such
a game into Coduno, the CatcoderGameImportService now looks for such a resource-file in
the xml file of the challenge description. If found, the file is uploaded into a bucket and the
link is provided in the challenge-instructions.
Also, the ChallengeTemplateController respectively Service returns a String instead of a UUID
since this is needed for the app to redirect to the correct page.
As uploading files is the most time-consuming part when importing a challenge, the currently
uploaded filename is now logged.