CatCoder: handle multiple resource files
In addition to a *challenge*-resource file, each *task/level* of a challenge can now
have its own resource file, which can be downloaded through a link in the task-
instructions.

There are existing catcoder-challenges that have one resource file that is used across
levels/tasks. Usually, this was distributed in 'level one'. So, if a challenge is imported
which has a resource file in 'level one' only the file is added to the challenge-
instructions (instead of task-instructions)