Task download restriction
When participating in a challenge, there are several tasks/levels. New tasks can only be shown
respectively downloaded when the previous challenge was submitted and produced the
correct results.
Previously, this was just handled on the client side and users could have downloaded the task
description of the next level by calling the corresponding API. This is not possible anymore as
the permission to download the file is now checked server-side.
This is done by setting a session attribute named "CURRENT_CHALLENGE" which is set when
the challenge is viewed and allows the server to quickly check in what level the user is.